rename input/ to keyboard/
